# docs/api/openapi.yaml
openapi: 3.1.2
info:
  title: Kontrak HTTP Ruvana
  version: 1.0.0
  description: |
    Kontrak HTTP untuk aplikasi reservasi dan pelaporan fasilitas kampus Ruvana.
    Semua tanggal kalender bisnis dimaknai dalam zona waktu Asia/Jakarta.
    Semua instant pada respons menggunakan UTC dan wajib berakhiran Z. Operasi
    yang memakai sesi memuat user terbaru dan hanya menerima akun berstatus
     ACTIVE. Akun non-ACTIVE selalu menghasilkan 401 generik; hanya principal
     ACTIVE dengan role yang tidak diterima yang menghasilkan 403. Authorization
     selalu dievaluasi sebelum lookup atau replay idempotensi. Untuk operasi yang
     mendeklarasikan Idempotency-Key, urutannya wajib ketat: authenticate ->
     verifikasi akun ACTIVE dan otorisasi role -> verifikasi ownership bila
     berlaku -> validasi Origin -> lookup/storage idempotensi -> operasi bisnis.
     Field lifecycle nullable pada Reservation, Report, dan
    Analytics selalu dikirim dengan nilai null ketika belum berlaku; field lain
    tidak boleh berganti antara omission dan null. Request field opsional boleh
    dihilangkan; nilai null hanya diterima bila schema menyatakannya dan berarti
    mengosongkan nilai (misalnya FacilityUpdateRequest.deskripsi). Origin wajib
    hadir tepat satu kali dan dibandingkan dengan scheme+host+port secara exact
     terhadap allowlist konfigurasi; missing, malformed, null, dan multiple Origin
    values selalu ditolak. Idempotency berlaku hanya untuk operasi yang
...
  contact:
...
servers:
...
security:
  - cookieAuth: []
paths:
...
  /api/admin/facilities/{facilityId}:
    get:
      operationId: getAdminFacility
      tags: [Administration]
      summary: Menampilkan detail fasilitas admin
      description: Hanya role=admin dengan status=ACTIVE yang diterima; akun non-ACTIVE menghasilkan 401 generik, sedangkan principal ACTIVE dengan role berbeda menghasilkan 403. Admin dapat melihat status dan metadata fasilitas termasuk INACTIVE.
      x-requirement-ids: [FAC-05, IAM-03]
      security: [{ cookieAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/FacilityId'
      responses:
        '400': { $ref: '#/components/responses/BadRequest' }
        '200':
          description: Detail fasilitas administrasi.
          headers:
            Cache-Control: { $ref: '#/components/headers/NoStore' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Facility' }
              example: { id: 1, nama: RK-101, tipe: ruang_kelas, lokasi: Gedung A Lt.1, kapasitas: 40, deskripsi: Ruang kelas standar ber-AC, status: ACTIVE, statusChangedAt: null, statusChangedBy: null }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalServerError' }
    patch:
      operationId: updateFacility
      tags: [Administration]
      summary: Memperbarui atau menonaktifkan fasilitas
      description: >-
          Hanya role admin dengan status ACTIVE yang diterima; akun dengan status selain
          ACTIVE selalu menghasilkan 401 generik, sedangkan principal ACTIVE dengan role
          berbeda menghasilkan 403. Matriks transisi admin yang valid tepat:
          ACTIVE -> UNDER_MAINTENANCE, UNDER_MAINTENANCE -> ACTIVE, ACTIVE -> INACTIVE,
          INACTIVE -> ACTIVE, dan UNDER_MAINTENANCE -> INACTIVE; transisi lain ditolak.
          Petugas hanya boleh ACTIVE <-> UNDER_MAINTENANCE dan tidak pernah boleh
          menangani INACTIVE. Hanya destination UNDER_MAINTENANCE yang memanggil
          REP-04/RES-09 dengan transaksi atomik yang sama: statusChangedBy adalah admin
          terbaru, statusChangedAt adalah satu timestamp server bersama, tepat reservasi
          APPROVED pada facility ini dengan startsAt > statusChangedAt dibatalkan dengan
          alasan otomatis yang terlihat user. Retry idempotent, reservation lampau/facility
          lain terlindungi, rollback berlaku bila efek gagal, dan destination ACTIVE
           menghitung ulang availability dari APPROVED terkini. Untuk setiap perubahan
           status, sebelum mengembalikan 200 server wajib invalidate seluruh representasi
           public list/detail/availability yang terdampak; client wajib refetch representasi
           tersebut. Minimal satu field wajib dikirim dan nama tetap unik; penghapusan
           permanen dilakukan lewat operasi delete yang menolak fasilitas beriwayat.
      x-requirement-ids: [FAC-05, FAC-04, REP-04, RES-09, IAM-03]
      security: [{ cookieAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/FacilityId'
        - $ref: '#/components/parameters/Origin'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FacilityUpdateRequest' }
            example: { deskripsi: Ruang kelas standar dengan proyektor baru. }
      responses:
        '200':
          description: >-
            Fasilitas terbaru; seluruh representasi public list/detail/availability
            yang terdampak telah di-invalidate sebelum 200 dikembalikan bila status
            berubah.
          headers:
            Cache-Control: { $ref: '#/components/headers/NoStore' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Facility' }
              example: { id: 1, nama: RK-101, tipe: ruang_kelas, lokasi: Gedung A Lt.1, kapasitas: 40, deskripsi: Ruang kelas standar dengan proyektor baru., status: ACTIVE, statusChangedAt: null, statusChangedBy: null }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/MutationForbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/FacilityMutationConflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '500': { $ref: '#/components/responses/InternalServerError' }
    delete:
      operationId: deleteFacility
      tags: [Administration]
      summary: Menghapus permanen fasilitas tanpa riwayat
      description: >-
        Hanya role admin dengan status ACTIVE. Hapus permanen hanya diizinkan bila
        fasilitas belum memiliki riwayat (reservasi atau laporan); bila ada, server
        menolak dengan 409 dan menyarankan penonaktifan (status INACTIVE). Foto
        unggahan fasilitas ikut dibersihkan dan representasi publik terdampak
        di-invalidate sebelum 204.
      x-requirement-ids: [FAC-05, IAM-03]
      security: [{ cookieAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/FacilityId'
        - $ref: '#/components/parameters/Origin'
      responses:
        '204':
          description: Fasilitas dihapus permanen.
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/MutationForbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/FacilityMutationConflict' }
        '500': { $ref: '#/components/responses/InternalServerError' }
...
components:
  securitySchemes:
    cronBearer:
      type: http
      scheme: bearer
    cookieAuth:
      type: apiKey
      in: cookie
      name: ruvana_session
      description: Cookie sesi persisten HttpOnly, SameSite=Lax, Secure di production, Path=/, dengan expiry server dan lifetime bounded yang dikonfigurasi implementasi (durasi tidak ditetapkan oleh kontrak ini).
  parameters:
    Origin:
      name: Origin
      in: header
      required: true
      description: >-
        Header harus hadir tepat satu kali. Server menolak missing, malformed,
        nilai literal null, dan multiple Origin values dengan 403
        CSRF_ORIGIN_REJECTED. Nilai yang valid dibandingkan tanpa normalisasi
        terhadap allowlist menggunakan exact scheme+host+port; path, query,
        fragment, dan wildcard tidak diterima.
      schema: { type: string, format: uri, maxLength: 2048 }
      example: https://app.example.invalid
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: >-
        UUID untuk operasi yang mendeklarasikan header ini, dengan identity
        authenticated principal + HTTP method/operation ID + canonical path
        parameters + relevant query + canonical body; retensi 24 jam. Urutan
        wajib adalah authenticate -> verifikasi akun ACTIVE dan otorisasi role
        -> verifikasi ownership bila berlaku -> validasi Origin ->
        idempotency lookup/storage -> operasi bisnis. Key/payload sama menunggu
        request pertama lalu mengembalikan status/body yang sama; payload berbeda
        menghasilkan 409 IDEMPOTENCY_KEY_REUSED. Request bersamaan dengan identity
        sama hanya menjalankan satu efek. Hanya hasil deterministik yang
        didokumentasikan oleh operasi yang boleh dipersist atau di-replay; 401,
        403, dan Origin yang ditolak tidak pernah dipersist atau di-replay, serta
        tidak mengonsumsi key. Status 200, 201, 204, dan 4xx deterministik yang
        didokumentasikan dapat dipersist; 5xx, 429, 502, dan 503 tidak dipersist.
        Canonical path binding mencegah replay lintas resource.
      schema: { type: string, format: uuid }
      example: 2c1f4a3e-7e71-4d5c-9f1c-0f0d9e5f1a11
...
    FacilityId:
      name: facilityId
      in: path
      required: true
      description: Identifier fasilitas positif.
      schema: { type: integer, minimum: 1 }
      example: 1
...
  headers:
...
    NoStore:
      description: Respons tidak boleh disimpan oleh browser, proxy, atau CDN.
      schema:
        type: string
        const: no-store
...
  responses:
    BadRequest:
      description: JSON atau parameter query/path malformed.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: https://app.example.invalid/problems/bad-request
            title: Permintaan tidak valid
            status: 400
            detail: JSON atau parameter tidak dapat diproses.
            instance: /api/resource
            code: BAD_REQUEST
    Unauthorized:
      description: Sesi hilang, tidak valid, kedaluwarsa, atau user tidak ACTIVE. Respons tidak membocorkan penyebab akun.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: https://app.example.invalid/problems/unauthorized
            title: Autentikasi diperlukan
            status: 401
            detail: Sesi tidak valid atau akun tidak aktif.
            instance: /api/protected
            code: UNAUTHORIZED
...
    Forbidden:
      description: ACTIVE principal dengan role yang tidak diterima untuk operasi ini menerima 403 non-CSRF dengan code FORBIDDEN.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ForbiddenProblem' }
          example:
            type: https://app.example.invalid/problems/forbidden
            title: Akses ditolak
            status: 403
            detail: Role akun tidak berwenang menjalankan operasi ini.
            instance: /api/protected
            code: FORBIDDEN
    MutationForbidden:
      description: >-
        Semua mutation 403 mencakup dua penyebab yang sengaja dibedakan oleh
         code: FORBIDDEN untuk role atau kewenangan yang tidak mencukupi,
        dan CSRF_ORIGIN_REJECTED untuk Origin missing, malformed, null, multiple,
        atau tidak cocok. Origin comparison memakai exact scheme+host+port
        terhadap allowlist.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/MutationForbiddenProblem' }
          examples:
            forbidden:
              value:
                type: https://app.example.invalid/problems/forbidden
                title: Akses ditolak
                status: 403
                detail: Role akun tidak berwenang menjalankan operasi ini.
                instance: /api/protected
                code: FORBIDDEN
            csrf:
              value:
                type: https://app.example.invalid/problems/csrf-origin-rejected
                title: Permintaan ditolak
                status: 403
                detail: Origin permintaan tidak diizinkan.
                instance: /api/protected
                code: CSRF_ORIGIN_REJECTED
    NotFound:
      description: Resource tidak ada atau ownership/private resource dimasking.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: https://app.example.invalid/problems/not-found
            title: Resource tidak ditemukan
            status: 404
            detail: Resource yang diminta tidak ditemukan.
            instance: /api/resource/1
            code: NOT_FOUND
...
    UnprocessableEntity:
      description: Input valid secara bentuk tetapi melanggar validasi field atau aturan bisnis.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: https://app.example.invalid/problems/validation-failed
            title: Validasi gagal
            status: 422
            detail: Satu atau lebih field tidak memenuhi aturan.
            instance: /api/resource
            code: VALIDATION_FAILED
            errors:
              - field: startTime
                code: SLOT_ALIGNMENT_INVALID
                message: Waktu harus berada pada kelipatan 30 menit.
...
    FacilityMutationConflict:
      description: Nama fasilitas, transisi status, atau Idempotency-Key menimbulkan konflik.
      content:
        application/problem+json:
          schema:
            anyOf:
              - $ref: '#/components/schemas/FacilityMutationProblem'
              - $ref: '#/components/schemas/FacilityTransitionProblem'
              - $ref: '#/components/schemas/IdempotencyConflictProblem'
          examples:
            name:
              value:
                type: https://app.example.invalid/problems/facility-name-already-used
                title: Nama fasilitas sudah digunakan
                status: 409
                detail: Nama fasilitas RK-101 harus unik.
                instance: /api/admin/facilities
                code: FACILITY_NAME_ALREADY_USED
            transition:
              value:
                type: https://app.example.invalid/problems/invalid-facility-transition
                title: Transisi fasilitas tidak valid
                status: 409
                detail: Transisi status fasilitas tidak diizinkan dari status saat ini.
                instance: /api/admin/facilities/1
                code: INVALID_FACILITY_TRANSITION
            idempotency:
              value:
                type: https://app.example.invalid/problems/idempotency-key-reused
                title: Idempotensi tidak cocok
                status: 409
                detail: Idempotency-Key telah digunakan untuk payload berbeda pada identity yang sama.
                instance: /api/admin/facilities/1
                code: IDEMPOTENCY_KEY_REUSED
    InternalServerError:
      description: Kesalahan tak terduga; detail internal dicatat server dan tidak dikirim ke client.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: https://app.example.invalid/problems/internal-error
            title: Kesalahan internal
            status: 500
            detail: Terjadi kesalahan yang tidak terduga.
            instance: /api/resource
            code: INTERNAL_ERROR
  schemas:
...
    FacilityType:
      type: string
      enum: [ruang_kelas, aula, laboratorium, alat, lapangan]
    FacilityStatus:
      type: string
      enum: [ACTIVE, UNDER_MAINTENANCE, INACTIVE]
...
    UtcInstant:
      type: string
      format: date-time
      pattern: 'Z$'
      description: Instant ISO 8601 canonical UTC; wajib memiliki trailing Z, bukan offset lokal atau offset numerik. Untuk expiry reservasi, server menetapkan EXPIRED ketika serverNow >= startsAt.
...
    StaffUser:
      type: object
      additionalProperties: false
      required: [id, nama, role]
      properties:
        id: { type: integer, minimum: 1, readOnly: true }
        nama: { type: string, minLength: 1, maxLength: 100, readOnly: true }
        role: { type: string, enum: [petugas, admin], readOnly: true }
...
    FacilityUpdateRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        nama: { type: string, minLength: 1, maxLength: 100 }
        tipe: { $ref: '#/components/schemas/FacilityType' }
        lokasi: { type: string, minLength: 1, maxLength: 200 }
        kapasitas: { type: integer, minimum: 1 }
        deskripsi: { type: [string, 'null'], maxLength: 2000 }
        status: { $ref: '#/components/schemas/FacilityStatus' }
        fotoPathname: { type: [string, 'null'], maxLength: 300, description: String mengganti foto, null menghapus, tidak dikirim membiarkan tidak berubah. }
        fotoType: { type: string, enum: [image/jpeg, image/png, image/webp] }
        fotoSize: { type: integer, minimum: 1, maximum: 5242880 }
...
    Facility:
      type: object
      additionalProperties: false
      required: [id, nama, tipe, lokasi, kapasitas, deskripsi, status, fotoUrl, statusChangedAt, statusChangedBy]
      properties:
        id: { type: integer, minimum: 1, readOnly: true }
        nama: { type: string, minLength: 1, maxLength: 100 }
        tipe: { $ref: '#/components/schemas/FacilityType' }
        lokasi: { type: string, minLength: 1, maxLength: 200 }
        kapasitas: { type: integer, minimum: 1 }
        deskripsi: { type: [string, 'null'], maxLength: 2000 }
        fotoUrl: { type: [string, 'null'], readOnly: true, description: URL same-origin foto unggahan; null bila memakai placeholder statis. }
        status: { $ref: '#/components/schemas/FacilityStatus' }
        statusChangedAt:
          anyOf:
            - $ref: '#/components/schemas/UtcInstant'
            - type: 'null'
          readOnly: true
          description: Provenance status; selalu dikirim pada schema operasional/admin dan bernilai null bila belum tersedia.
        statusChangedBy:
          anyOf:
            - $ref: '#/components/schemas/StaffUser'
            - type: 'null'
          readOnly: true
          description: Aktor provenance status; selalu dikirim pada schema operasional/admin dan bernilai null bila belum tersedia.
...
    Problem:
      allOf:
        - $ref: '#/components/schemas/ProblemBase'
      unevaluatedProperties: false
      description: Closed generic RFC 9457 problem detail.
    ProblemBase:
      type: object
      required: [type, title, status, detail, instance, code]
      properties:
        type: { type: string, format: uri, description: URI problem type yang stabil. }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        detail: { type: string }
        instance: { type: string, description: 'Identifier instance request, bukan rahasia.' }
        code: { type: string, description: Kode error machine-readable. }
        errors:
          type: array
          items: { $ref: '#/components/schemas/ProblemFieldError' }
    MutationForbiddenProblem:
      allOf:
        - $ref: '#/components/schemas/ProblemBase'
        - type: object
          required: [status, code]
          properties:
            status: { type: integer, const: 403 }
            code: { type: string, enum: [FORBIDDEN, CSRF_ORIGIN_REJECTED] }
      unevaluatedProperties: false
    ForbiddenProblem:
      allOf:
        - $ref: '#/components/schemas/ProblemBase'
        - type: object
          required: [status, code]
          properties:
            status: { type: integer, const: 403 }
            code: { type: string, const: FORBIDDEN }
      unevaluatedProperties: false
    IdempotencyConflictProblem:
      allOf:
        - $ref: '#/components/schemas/ProblemBase'
        - type: object
          required: [status, code]
          properties:
            status: { type: integer, const: 409 }
            code: { type: string, const: IDEMPOTENCY_KEY_REUSED }
      unevaluatedProperties: false
    ProblemFieldError:
      type: object
      additionalProperties: false
      required: [field, code, message]
      properties:
        field: { type: string }
        code: { type: string }
        message: { type: string }
...
    FacilityTransitionProblem:
      allOf:
        - $ref: '#/components/schemas/ProblemBase'
        - type: object
          required: [status, code]
          properties:
            status: { type: integer, const: 409 }
            code: { type: string, const: INVALID_FACILITY_TRANSITION }
      unevaluatedProperties: false
    FacilityMutationProblem:
      allOf:
        - $ref: '#/components/schemas/ProblemBase'
        - type: object
          required: [status, code]
          properties:
            status: { type: integer, const: 409 }
            code: { type: string, const: FACILITY_NAME_ALREADY_USED }
      unevaluatedProperties: false